requirements.txt
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.0
fsspec==2024.2.0
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pandas==2.2.2
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
typing_extensions==4.9.0
tzdata==2024.1
wheel==0.43.0
